AgentInterfaceTest/requirements.txt
requests==2.25.1
yagmail==0.14.256
PyYAML==5.4.1
ddt==1.4.2
PyMySQL==1.0.2
loguru==0.5.3
redis~=4.2.2
DBUtils~=3.0.2